requirements.txt
pillow
pynetdicom
pydicom
numpy
tweepy